Revert "Add serialisation of agents"
This reverts commit 04b45952e722d21c0ef9370f7c5902694b1dbe2b.

Put simply, agents cannot be serialised in its current state.

This is because Agent maintains a Rc to Subculture and Neighbourhood, and
when deserialised serde creates a new Rc for Subculture and Neighbourhood
essentially duplicating the data, creating higher memory requirements, and
preventing interventions as Subculture and Neighbourhood cannot be updated.